+++ include/endstone/metrics/simple_bar_chart.h
@@ -17,8 +17,8 @@
 #include <functional>
 #include <optional>
 #include <string>
+#include <unordered_map>
 #include <utility>
-#include <vector>
 
 #include "endstone/metrics/custom_chart.h"
 
@@ -27,7 +27,7 @@ namespace endstone {
 /** A bStats bar chart with a single bar per category. */
 class SimpleBarChart : public CustomChart {
 public:
-    using Callback = std::function<std::optional<StringValues>()>;
+    using Callback = std::function<std::optional<std::unordered_map<std::string, int>>()>;
 
     SimpleBarChart(std::string chart_id, Callback get_values)
         : CustomChart(std::move(chart_id)), get_values_(std::move(get_values))