+++ src/endstone/core/level/level.cpp
@@ -31,31 +31,34 @@
 #include "endstone/core/game_rule.h"
 #include "endstone/core/inventory/recipe_data.h"
 #include "endstone/core/level/dimension.h"
+#include "endstone/event/level/dimension_load_event.h"
 #include "endstone/level/dimension.h"
+#include "endstone/plugin/plugin_manager.h"
 
 namespace endstone::core {
 
-EndstoneLevel::EndstoneLevel(::Level &level) : server_(EndstoneServer::getInstance()), level_(level)
+EndstoneLevel::EndstoneLevel(::Level &level) : server_(EndstoneServer::getInstance()), level_(level) {}
+
+void EndstoneLevel::loadDimensions()
 {
-    // Load all vanilla dimensions on start up
-    level.getOrCreateDimension(VanillaDimensions::Overworld);
-    level.getOrCreateDimension(VanillaDimensions::Nether);
-    level.getOrCreateDimension(VanillaDimensions::TheEnd);
-    auto add_dimension = [this](::Dimension &dimension) {
-        dimensions_[dimension.getDimensionId().value] =
-            std::make_shared<EndstoneDimension>(dimension.getWeakRef(), *this);
-    };
-    level.forEachDimension([&](::Dimension &dimension) {
+    level_.getDimensionManager().getOnNewDimensionCreatedConnector().connect(
+        [&](::Dimension &dimension) { addDimension(dimension); }, Bedrock::PubSub::ConnectPosition::AtBack, nullptr);
+    level_.forEachDimension([&](::Dimension &dimension) {
         addDimension(dimension);
         return true;
     });
-    level.getDimensionManager().getOnNewDimensionCreatedConnector().connect(
-        [&](::Dimension &dimension) { addDimension(dimension); }, Bedrock::PubSub::ConnectPosition::AtBack, nullptr);
+    // Load all vanilla dimensions on start up
+    level_.getOrCreateDimension(VanillaDimensions::Overworld);
+    level_.getOrCreateDimension(VanillaDimensions::Nether);
+    level_.getOrCreateDimension(VanillaDimensions::TheEnd);
 }
 
 void EndstoneLevel::addDimension(::Dimension &dimension)
 {
-    dimensions_[dimension.getDimensionId().value] = std::make_shared<EndstoneDimension>(dimension.getWeakRef(), *this);
+    const auto handle = std::make_shared<EndstoneDimension>(dimension.getWeakRef(), *this);
+    dimensions_[dimension.getDimensionId().value] = handle;
+    DimensionLoadEvent event{handle};
+    server_.getPluginManager().callEvent(event);
 }
 
 std::string EndstoneLevel::getName() const