+++ tests/plugin/src/endstone_test/listeners/level_event_listener.py
@@ -1,4 +1,9 @@
-from endstone.event import ChunkLoadEvent, ChunkUnloadEvent, event_handler
+from endstone.event import (
+    ChunkLoadEvent,
+    ChunkUnloadEvent,
+    DimensionLoadEvent,
+    event_handler,
+)
 
 from .event_listener import EventListener
 
@@ -9,7 +14,6 @@ class LevelEventListener(EventListener):
         self.record(
             event,
             f"{event.chunk} is loaded",
-            always_log=True,
             chunk_xz=(event.chunk.x, event.chunk.z),
             dimension=str(event.dimension.id),
         )
@@ -19,7 +23,15 @@ class LevelEventListener(EventListener):
         self.record(
             event,
             f"{event.chunk} is unloaded",
-            always_log=True,
             chunk_xz=(event.chunk.x, event.chunk.z),
+            dimension=str(event.dimension.id),
+        )
+
+    @event_handler
+    def on_dimension_load(self, event: DimensionLoadEvent):
+        self.record(
+            event,
+            f"{event.dimension.id} is loaded",
+            always_log=True,
             dimension=str(event.dimension.id),
         )